abtracted base global into HOC
